Congo twice in the list #6
As i understand Congo and The democratic republic of the Congo it two
different countries. Congo has another name - Republic of the Congo.
Also i`ve moved The Democratic Republic of the Congo before Congo, The
Democratic Republic of The. So now in the list of countries first goes
Congo and then The Democratic Republic of the Congo.